hairsync/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtpd==1.4.5
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
atpublic==4.0
attrs==23.2.0
Authlib==1.3.0
bcrypt==4.1.2
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distlib==0.3.8
-e git+https://github.com/fae713/django.git@921670c6943e9c532137b7d164885f2d3ab436b8#egg=Django
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
docutils==0.20.1
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
geoip2==4.8.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
maxminddb==2.6.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
pylibmc==1.6.3
pymemcache==4.0.0
PySocks==1.7.1
pywatchman==2.0.0
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tblib==3.0.0
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.9.4
